requirements.txt
openai
pdfplumber
python-dotenv
requests
lxml
